requirements.txt
numpy==1.20.3
opencv_python_headless==4.7.0.68
Pillow==9.4.0
torch==1.13.0
torchvision==0.13.1
